Restore Egg* functions for dealing with virtual modifiers
Everything forgiven, Egg*! Please come back..

Virtual modifiers are SUPER, META etc and we need to map them back and
forth between GTK+ and X. Egg* does that, and we have to wait for GTK+
2.20 until:

    gdk_keymap_add_virtual_modifiers
    gdk_keymap_map_virtual_modifiers

The virtualization mapping was modified to correspond to the GTK+
version.

The code is now definitely beta quality, there may still be many bugs.